+++ teuthology/misc.py
@@ -15,6 +15,9 @@ import yaml
 import json
 import re
 import pprint
+import datetime
+
+from tarfile import ReadError
 
 from teuthology.util.compat import urljoin, urlopen, HTTPError
 
@@ -31,7 +34,6 @@ from teuthology.orchestra.opsys import DEFAULT_OS_VERSION
 
 log = logging.getLogger(__name__)
 
-import datetime
 stamp = datetime.datetime.now().strftime("%y%m%d%H%M")
 
 is_arm = lambda x: x.startswith('tala') or x.startswith(
@@ -1355,8 +1357,52 @@ def add_remote_path(ctx, local_dir, remote_dir):
     with open(os.path.join(ctx.archive, 'info.yaml'), 'r+') as info_file:
         info_yaml = yaml.safe_load(info_file)
         info_file.seek(0)
-        if 'archive' not in info_yaml:
-            info_yaml['archive'] = {local_dir: remote_dir}
-        else:
+        if 'archive' in info_yaml:
             info_yaml['archive'][local_dir] = remote_dir
+        else:
+            info_yaml['archive'] = {local_dir: remote_dir}
         yaml.safe_dump(info_yaml, info_file, default_flow_style=False)
+
+
+def archive_logs(ctx, remote_path, log_path):
+    """
+    Archive directories from all nodes in a cliuster. It pulls all files in
+    remote_path dir to job's archive dir under log_path dir.
+    """
+    path = os.path.join(ctx.archive, 'remote')
+    os.mkdirs(path, exist_ok=True)
+    for remote in ctx.cluster.remotes.keys():
+        sub = os.path.join(path, remote.shortname)
+        os.mkdirs(sub, exist_ok=True)
+        try:
+            pull_directory(remote, remote_path, os.path.join(sub, log_path))
+        except ReadError:
+            pass
+
+
+def compress_logs(ctx, remote_dir):
+    """
+    Compress all files in remote_dir from all nodes in a cluster.
+    """
+    log.info('Compressing logs...')
+    run.wait(
+        ctx.cluster.run(
+            args=[
+                'sudo',
+                'find',
+                remote_dir,
+                '-name',
+                '*.log',
+                '-print0',
+                run.Raw('|'),
+                'sudo',
+                'xargs',
+                '-0',
+                '--no-run-if-empty',
+                '--',
+                'gzip',
+                '--',
+            ],
+            wait=False,
+        ),
+    )
